RxSamples.playground
* Errors fixes in playgrounds
* Updated playgrounds to Xcode 7 paged playground
