--- Playgrounds/Introduction.playground/Sources/SupportCode.swift
@@ -1,6 +1,0 @@
-
-
-public func example(description: String, action: () -> ()) {
-    print("\n--- \(description) example ---")
-    action()
-}
