--- Playgrounds/ObservablesOperators/Observables+Error Handling.playground/contents.xcplayground
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='5.0' target-platform='osx' requires-full-environment='true' display-mode='rendered'>
-    <timeline fileName='timeline.xctimeline'/>
-</playground>
