+++ RxSamples.playground/Pages/Index.xcplaygroundpage/Contents.swift
@@ -1,0 +1,28 @@
+
+/*:
+> # IMPORTANT: To use `RxSamples`, please:
+
+1. Open `Rx.xcworkspace`
+2. Build `RxSwift-OSX` scheme
+3. And then open `RxSamples` in `Rx.xcworkspace` tree view.
+4. View the console to view log strings. Choose `View > Show Debug Area`
+*/
+
+/*:
+## Index:
+
+1. [Introduction](Introduction)
+2. [Subjects](Subjects)
+3. [Creating observables](Creating_observables)
+4. [Transforming Observables](Transforming_Observables)
+5. [Filtering Observables](Filtering_Observables)
+6. [Combining Observables](Combining_Observables)
+7. [Error Handling Operators](Error_Handling_Operators)
+8. [Observable Utility Operators](Observable_Utility_Operators)
+9. [Conditional and Boolean Operators](Conditional_and_Boolean_Operators)
+10. [Mathematical and Aggregate Operators](Mathematical_and_Aggregate_Operators)
+11. [Connectable Observable Operators](Connectable_Observable_Operators)
+
+*/
+
+//: [Index](Index) - [Next >>](@next)
